NETFLIX-BUILD: [LA-1841] Ignore Jars and S3 files when preparing SparkR.
(cherry picked from commit 9ae54b40973bfdd14aff54cb81eac67c10de176f)

Conflicts:
	core/src/main/scala/org/apache/spark/deploy/RPackageUtils.scala

+++ core/src/main/scala/org/apache/spark/deploy/RPackageUtils.scala
@@ -176,37 +176,45 @@ private[deploy] object RPackageUtils extends Logging {
       printStream: PrintStream = null,
       verbose: Boolean = false): Unit = {
     jars.split(",").foreach { jarPath =>
-      val file = new File(Utils.resolveURI(jarPath))
-      if (file.exists()) {
-        val jar = new JarFile(file)
-        Utils.tryWithSafeFinally {
-          if (checkManifestForR(jar)) {
-            print(s"$file contains R source code. Now installing package.", printStream, Level.INFO)
-            val rSource = extractRFolder(jar, printStream, verbose)
-            if (RUtils.rPackages.isEmpty) {
-              RUtils.rPackages = Some(Utils.createTempDir().getAbsolutePath)
-            }
-            try {
-              if (!rPackageBuilder(rSource, printStream, verbose, RUtils.rPackages.get)) {
-                print(s"ERROR: Failed to build R package in $file.", printStream)
-                print(RJarDoc, printStream)
+      val uri = Utils.resolveURI(jarPath)
+      // only process local jar files for Spark source code
+      if (uri.getScheme == "file" && new File(uri.getPath).getName.endsWith("jar")) {
+        val file = new File(uri)
+        if (file.exists()) {
+          val jar = new JarFile(file)
+          Utils.tryWithSafeFinally {
+            if (checkManifestForR(jar)) {
+              print(s"$file contains R source code. Now installing package.", printStream,
+                Level.INFO)
+              val rSource = extractRFolder(jar, printStream, verbose)
+              if (RUtils.rPackages.isEmpty) {
+                RUtils.rPackages = Some(Utils.createTempDir().getAbsolutePath)
               }
-            } finally {
-              // clean up
-              if (!rSource.delete()) {
-                logWarning(s"Error deleting ${rSource.getPath()}")
+              try {
+                if (!rPackageBuilder(rSource, printStream, verbose, RUtils.rPackages.get)) {
+                  print(s"ERROR: Failed to build R package in $file.", printStream)
+                  print(RJarDoc, printStream)
+                }
+              } finally {
+                // clean up
+                if (!rSource.delete()) {
+                  logWarning(s"Error deleting ${rSource.getPath()}")
+                }
+              }
+            } else {
+              if (verbose) {
+                print(s"$file doesn't contain R source code, skipping...", printStream)
               }
             }
-          } else {
-            if (verbose) {
-              print(s"$file doesn't contain R source code, skipping...", printStream)
-            }
+          } {
+            jar.close()
           }
-        } {
-          jar.close()
+        } else {
+          print(s"WARN: $file resolved as dependency, but not found.", printStream, Level.WARNING)
         }
       } else {
-        print(s"WARN: $file resolved as dependency, but not found.", printStream, Level.WARNING)
+        print(s"WARN: Not processing $uri for R source code (not a local Jar)",
+          printStream, Level.WARNING)
       }
     }
   }
